requirements.txt
opencv-python==4.7.0.72
Pillow==9.4.0
scikit-learn==1.3.0
pandas==2.0.3
matplotlib==3.8.0
